Employee Rest API using Hibernates

+++ section4/EmployeeRestAPI/src/main/java/com/employees/EmployeeRestAPI/EmployeeRESTControlleer.java
@@ -1,6 +1,7 @@
 package com.employees.EmployeeRestAPI;
 
 import com.employees.EmployeeRestAPI.entity.Employee;
+import com.employees.EmployeeRestAPI.exception.EmployeeNotFoundException;
 import com.employees.EmployeeRestAPI.service.DaoService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -23,6 +24,9 @@ public class EmployeeRESTControlleer {
     }
     @GetMapping("/employee/{id}")
     public Employee findEmployeebyID(@PathVariable int id){
+        if(!daoService.findAll().contains(id)|| id < 0){
+            throw new EmployeeNotFoundException("This current Employee ID is not available");
+        }
         return daoService.employeeByID(id);
     }
     @PostMapping("/employee")
@@ -33,6 +37,9 @@ public class EmployeeRESTControlleer {
     }
     @DeleteMapping("/employee/{id}")
     public Employee deleteEmployee(@PathVariable int id){
+        if(!daoService.findAll().contains(id)|| id < 0){
+            throw new EmployeeNotFoundException("This current Employee ID is not available");
+        }
         return daoService.deleteEmployee(id);
     }
    @PutMapping("/employee")
